@@ -2,12 +2,12 @@ package com.okta.rest.quarkus;
 
 import io.quarkus.security.Authenticated;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.SecurityContext;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.SecurityContext;
 import java.security.Principal;
 
 @Path("/hello")
@@ -19,7 +19,7 @@ public class HelloResource {
     @Produces(MediaType.TEXT_PLAIN)
     public String hello(@Context SecurityContext context) {
         Principal userPrincipal = context.getUserPrincipal();
-        return "Hello, " + userPrincipal.getName() + "!";
+        return "Hello from Quarkus, " + userPrincipal.getName() + "!";
     }
 
 }
